GENERADOR DE NUMERO DE EXPEDIENTE

+++ resources/views/roles/index.blade.php
@@ -57,6 +57,7 @@
 										@endcan
 									</td>
 									<td width="10px">
+									@if($role->id > 4 )
 										@can('roles.destroy')
 											{!! Form::open(['route' => ['roles.destroy', $role->id],
 											'method' => 'DELETE']) !!}
@@ -65,6 +66,7 @@
 												</button>
 											{!! Form::close() !!}
 										@endcan
+									@endif
 									</td>
 								</tr>
 								@endforeach
